mdchatbot/requirements.txt
# ------------ INSTALLATION INSTRUCTIONS ------------

# ----------- FIRST INSTALL torch LIB -----------
#torch==2.6.0+cpu
#IF YOU ARE USING CPU VERSION OF TORCH THEN USE BELOW LINE
#SHOULD BE INSTALL BEFOR INSTALLING REQUIREMEENT.TXT BECAUSE IT IS NOT DIRECTLY AVAILABLE PyPI
#pip install torch==2.6.0 --index-url https://download.pytorch.org/whl/cpu
#BUT IF YOU ARE USING GPU VERSION OF TORCH THEN USE BELOW LINE
#torch==2.6.0+cu121

# Core Web Frameworks
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
Flask==3.1.1
flask-cors==6.0.0

# LangChain & Vector Search
langchain==0.3.23
langchain-community==0.3.21
langchain-huggingface==0.1.2
faiss-cpu==1.11.0

# LLM Embeddings
sentence-transformers==4.1.0
rank-bm25==0.2.2
transformers==4.51.3

scikit-learn==1.7.0rc1
numpy==2.2.6
scipy==1.15.3

# Google Gemini SDK
google-genai==1.15.0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
protobuf==5.29.4

# Utility Libraries
pypdf==5.4.0
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-core==2.33.1
typing-extensions==4.13.2
PyJWT==2.10.1

python-decouple